Synes BrevDataMapper er slitsam å finne fram i, så startar refaktorisering. Første ledd å trekkje ut datafetcher-klassa til eiga fil
